requirements.txt
python-dotenv
git+https://github.com/openai/swarm.git
